New gradients for stress-velocity formulation
New gradients for stress-velocity formulation (GRAD_FORM=2) using a
decomposed impedance matrix to introduce a self-adjoint forward
operator. Notice, that no numerical integration of the data residuals is
required. While this improves the robustness of the FWI against noise,
lower frequencies are required in the FWI workflow to mitigate
trade-offs between different parameter classes. The formulations
GRAD_FORM=3 and GRAD_FORM=4 are removed.